4.4.28      *  Windows XP: Praat files can finally again be opened by        double-clicking and by dragging them onto the Praat icon.      *  Scripting (Windows): removed a bug that caused Praat to crash if        the script window was closed when a file selector window was open.

+++ sys/enum.h
@@ -86,7 +86,7 @@ typedef struct enum_ANY { int _length; const char *_type, *zero, *_end; int _tra
 
 typedef signed char enum1;
 typedef signed short enum2;
-#define enumi(type,element)  (((int) & ((type *) 0) -> element - (int) & ((enum_ANY *) 0) -> zero) / sizeof (char *))
+#define enumi(type,element)  (((int) & ((type *) 0) -> element - (int) & ((enum_ANY *) 0) -> zero) / (int) sizeof (char *))
 
 #define enum_type(enumerated)  (((enum_ANY *) enumerated) -> _type)
